Risk Detector page UI attached with viewModel

+++ lib/data/repositories/risk_detector_repository.dart
@@ -2,5 +2,5 @@ import 'package:injectable/injectable.dart';
 
 @factoryMethod
 abstract class RiskDetectorRepository{
-  Future<String> fetchData(int age,int systolicBP,int diastolicBP,double bs,int bodyTemp,int heartRate);
+  Future<String> fetchData(int age,int systolicBP,int diastolicBP,double bs,double bodyTemp,int heartRate);
 }
